backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "SFMS Phase 1 Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asyncpg>=0.31.0",
    "boto3==1.42.46",
    "botocore==1.42.46",
    "click==8.3.1",
    "fastapi==0.115.11",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "hatch==1.14.0",
    "idna==3.10",
    "jwt==1.3.1",
    "passlib==1.7.4",
    "pydantic[email]==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.8.1",
    "python-multipart==0.0.20",
    "redis>=5.3.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.38",
    "starlette==0.46.1",
    "typing-extensions==4.12.2",
    "uvicorn==0.34.0",
    "bcrypt==4.2.1",
    "pyjwt==2.10.1",
]

[dependency-groups]
dev = [
    "ruff==0.9.10",
    "sqlfluff==3.3.1",
    "httpx>=0.28.1",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "C90", "I", "B", "UP", "D"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["app"]
